Updated imports and lists in the vacancy cards
