+++ jsconfig.json
@@ -1,0 +1,9 @@
+{
+    "compilerOptions": {
+      "paths": {
+        "@/*": [
+          "./src/*"
+        ]
+      }
+    }
+}
