+++ src/Components/Filters/Tablet/Tablet.jsx
@@ -1,7 +1,7 @@
 const Tablet = () => {   
     return(
           <div className="flex justify-between items-center bg-light-cyan-filter-tablets w-fit h-10">
-            <span className="text-desaturated-dark-cyan text-lg flex justify-center items-center p-2">Hello</span>
+            <span className="text-desaturated-dark-cyan text-lg flex justify-center items-center p-2">JavaScript</span>
             <button className="bg-desaturated-dark-cyan hover:bg-very-dark-cyan p-3"><img src="./Images/icon-remove.svg" alt="" /></button>
           </div>
     )
