+++ src/Components/Vacancys/Updates/Updates.jsx
@@ -1,0 +1,10 @@
+const Updates = () => {
+    return(
+        <ul className="flex justify-center items-center gap-3 font-bold text-light-cyan-background text-sm">
+            <li className="bg-desaturated-dark-cyan p-2 rounded-2xl">NEW!</li>
+            <li className="bg-very-dark-cyan p-2 rounded-2xl">FEATURED</li>
+        </ul>
+    )
+} 
+
+export default Updates;
