+++ src/Pages/Main/Main.jsx
@@ -1,13 +1,13 @@
-import Header from "../../Components/Header/Header";
-import Vacancy from "../../Components/Vacancy/Vacancy"
-import Filters from "../../Components/Filters/Filters/Filters";
+import Header from "@/Components/Header/Header";
+import Vacancy from "@/Components/Vacancys/Vacancy/Vacancy"
+import Filter from "@/Components/Filters/Filter/Filter";
 
 const Main = () => {
     return(
         <>
             <Header />
             
-            <Filters />
+            <Filter />
 
             <div className="flex justify-center items-center mt-6">
                 <Vacancy />
